+++ src/domain/usecases/ImportUseCase.ts
@@ -1,21 +1,39 @@
 import _ from "lodash";
 
+import {
+    GeeInterval,
+    GeeDataRepository,
+    GeeGeometry,
+    GeeDataSetId,
+    GeeDataFilters
+} from "../repositories/GeeDataRepository";
+import { OrgUnit } from "../entities/OrgUnit";
+import { DataValueSet, DataValue } from "../entities/DataValue";
+import OrgUnitRepository from "../repositories/OrgUnitRepository";
+import { GeeDataItem } from "../entities/GeeData";
+import { promiseMap } from "../utils";
+
 import { getImportCountString } from "../../webapp/utils/dhis2";
 import { Config } from "../../webapp/models/Config";
 import i18n from "../../webapp/locales";
-import { GeeDhis2 } from "../../webapp/models/GeeDhis2";
+import { DataImportData } from "../../webapp/models/Import";
 import { getAttributeMappings, getDataSetValue } from "../../webapp/utils/gee";
 import { buildPeriod, downloadFile } from "../../webapp/utils/import";
 import { D2Api } from "d2-api";
-import { DataImportData } from "../../webapp/models/Import";
-import { GeeInterval, GeeDataRepository } from "../repositories/GeeDataRepository";
-import { OrgUnit } from "../entities/OrgUnit";
-import { DataValueSet } from "../entities/DataValue";
-import OrgUnitRepository from "../repositories/OrgUnitRepository";
+
+type DataElementId = string;
+
+interface GetDataValueSetOptions<Band extends string> {
+    geeDataSetId: GeeDataSetId;
+    mapping: Record<Band, DataElementId>;
+    orgUnits: OrgUnit[];
+    interval: GeeInterval;
+    scale?: number;
+}
 
 //TODO: this use case is the old run method in old import model
-// little a little we are going to refactoring this use case 
-// creating adapters that invoke it until the usecase has not 
+// little a little we are going to refactoring this use case
+// creating adapters that invoke it until the usecase has not
 // webapp and infrastructure dependencies
 class ImportUseCase {
     constructor(private api: D2Api,
@@ -30,8 +48,6 @@ class ImportUseCase {
         let failures: string[] = [];
         let messages: string[] = [];
         try {
-
-            const geeDhis2 = GeeDhis2.init(this.api, this.geeDataRepository);
 
             const orgUnitIds = _.compact(data.selectedOUs.map(o => o.split("/").pop()));
             const orgUnits = await this.orgUnitRepository.getByIds(orgUnitIds);
@@ -49,7 +65,7 @@ class ImportUseCase {
             await Promise.all(
                 data.selectedMappings.map(async selectedMapping => {
                     try {
-                        const dataValueSet: DataValueSet = await geeDhis2.getDataValueSet({
+                        const dataValueSet: DataValueSet = await this.getDataValueSet({
                             ...baseImportConfig,
                             geeDataSetId: getDataSetValue(
                                 selectedMapping.geeImage,
@@ -104,6 +120,70 @@ class ImportUseCase {
                 success: false,
                 messages: messages,
                 failures: [...failures, i18n.t("Import config failed"), err],
+            };
+        }
+    }
+
+
+    private getGeometryFromOrgUnit(orgUnit: OrgUnit): GeeGeometry | undefined {
+        const coordinates = orgUnit.coordinates ? JSON.parse(orgUnit.coordinates) : null;
+        if (!coordinates) return;
+
+        switch (orgUnit.featureType) {
+            case "POINT":
+                return { type: "point", coordinates };
+            case "POLYGON":
+            case "MULTI_POLYGON":
+                return { type: "multi-polygon", polygonCoordinates: coordinates };
+            default:
+                return;
+        }
+    }
+
+    private async getDataValueSet<Band extends string>(
+        options: GetDataValueSetOptions<Band>
+    ): Promise<DataValueSet> {
+        const { geeDataRepository } = this;
+        const { geeDataSetId, orgUnits, mapping, interval, scale } = options;
+
+        const dataValuesList = await promiseMap(orgUnits, async (orgUnit) => {
+            const geometry = this.getGeometryFromOrgUnit(orgUnit);
+
+            if (!geometry) return [];
+
+            const options: GeeDataFilters<Band> = {
+                id: geeDataSetId,
+                bands: _.keys(mapping) as Band[],
+                geometry,
+                interval,
+                scale,
+            };
+
+            const geeData = await geeDataRepository.getData(options);
+
+            return _(geeData).map(item =>
+                this.mapGeeDataItemToDataValue(item, orgUnit.id, mapping)).compact().value()
+        });
+
+        return { dataValues: _.flatten(dataValuesList) };
+    }
+
+    private mapGeeDataItemToDataValue<Band extends string>(
+        item: GeeDataItem<Band>, orgUnitId: string,
+        mapping: Record<Band, DataElementId>): DataValue | undefined {
+
+        const { date, band, value } = item;
+        const dataElementId = mapping[band];
+
+        if (!dataElementId) {
+            console.error(`Band not found in mapping: ${band}`);
+            return;
+        } else {
+            return {
+                dataElement: dataElementId,
+                value: value.toFixed(18),
+                orgUnit: orgUnitId,
+                period: date.format("YYYYMMDD"), // Assume periodType="DAILY"
             };
         }
     }
