+++ src/domain/utils/index.ts
@@ -1,0 +1,11 @@
+/* Map sequentially over T[] with an async function and return array of mapped values */
+export function promiseMap<T, S>(inputValues: T[], mapper: (value: T) => Promise<S>): Promise<S[]> {
+    const reducer = (acc$: Promise<S[]>, inputValue: T): Promise<S[]> =>
+        acc$.then((acc: S[]) =>
+            mapper(inputValue).then(result => {
+                acc.push(result);
+                return acc;
+            })
+        );
+    return inputValues.reduce(reducer, Promise.resolve([]));
+}
